Plot a karyotype overview of the genome with the identified regions

Plots an overview of the genomic locations of the identified regions (see calculatePvalues) in a karyotype view. The coloring can be done either by significant regions according to their p-values, significant by adjusted p-values, or by annotated region if using matchGenes.

plotOverview(
  regions,
  annotation = NULL,
  type = "pval",
  significantCut = c(0.05, 0.1),
  ...
)

Arguments

regions
The $regions output from calculatePvalues.

annotation
The output from running matchGenes on the output from calculatePvalues. It is only required if type='annotation'.

type
Must be either pval, qval, fwer or annotation. It determines whether the plot coloring should be done according to significant p-values (<0.05), significant q-values (<0.10), significant FWER adjusted p-values (<0.05) or annotation regions.

significantCut
A vector of length two specifiying the cutoffs used to determine significance. The first element is used to determine significance for the p-values and the second element is used for the q-values.

...
Arguments passed to other methods and/or advanced arguments. Advanced arguments:

base_size

Base point size of the plot. This argument is passed to element_text (size argument).

areaRel

The relative size for the area label when type='pval' or type='qval'. Can be useful when making high resolution versions of these plots in devices like CairoPNG.

legend.position

This argument is passed to theme. From ggplot2: the position of legends. ('left', 'right', 'bottom', 'top', or two-element numeric vector).

Passed to extendedMapSeqlevels.

Value

A ggplot2 plot that is ready to be printed out. Tecnically it is a ggbio object.

See also

Examples